Handle Cloudflare Access URLs in the webview, not the browser.
We need to handle this in the webview, as certain cookies will be set after a
successful authentication. Cloudflare will only allow access to the underlying
server if these cookies are present. If the user would authenticate in the browser,
the cookies would be stored there and we would not be able to read them.