Improve usability when external url is behind a reverse proxy
Store cookies during the onboarding phase. This is already common
behavior for the remainder of the app. Usually a reverse proxy stores
a cookie, but this is lost after the onboarding which causes the app
to fail on next startup.

Don't launch the browser in the WebViewActivity when the url changes
while the app is not connected yet. Sometimes a reverse proxy wants to
show a different page before accepting the connection. This gives the
proxy the opportunity to do so. Otherwise the app can launch the
browser over 15 times.