Prevent socket response continuation being called multiple times (#3338) (#4616)
* Prevent socket response continuation being called multiple times (3338)

Coroutines continuations can only be invoke once, see CancellableContinuation. Use atomic boolean to ensure the continuation is called only once in a thread safe way.

* Update import ordering

* Add warning log message to coroutines continuation

Log a warning message so that it's more visible if a websocket response is invoked multiple times

* Set initial continuation invocation to false

The default value must be false as the continuation property is mutable. The value is always assigned after the object is created.