Fixup misleading nomenclature: s/frontend cache/client cache/g
This commit replaces every instance of a misleading reference to a "frontend
cache" where there is no such thing with a correct, however generalized,
reference to the "client cache."

In objective fact, a "frontend cache" is a term reserved for server-side caching
reverse proxies centralized on and attributable to a frontend itself. No such
resource exists in any Home Assistant deployment. This incorrect use of terms is
confusing and misleading, and therefore is herein replaced.

The only relevant cache is attributable to the client, whether it is a browser
or whether it is a dedicated platform client. This change categorically prevents
the mistaken apprehension that any server-side cache attributable to the
frontend requires resetting.